+++ R/dfactor.R
@@ -18,15 +18,15 @@
 #' dfactor(iris2, c("Color", "Species"))
 dfactor <- function (data1, vars, neat = TRUE, sizesort = TRUE) {
 
-  n           <- length(table(data1[, vars])) # Obtain levels length
   dataname    <- deparse(substitute(data1))   # Obtain data object name
+  n           <- length(table(data1[, vars])) # Obtain levels length
   dataset     <- neatify(dataname, n, neat)   # Replicate object name n times or neatly
   demographic <- neatify(vars, n, neat)       # Replicate demographic name n times or neatly
 
   if(neat){
-    dft <- data_frame_table(data1[, vars], prop = FALSE, perc = TRUE)
+    dft <- dft(data1[, vars], prop = FALSE, perc = TRUE)
   } else {
-    dft <- data_frame_table(data1[, vars], prop = TRUE, perc = TRUE)
+    dft <- dft(data1[, vars], prop = TRUE, perc = TRUE)
   }
 
   # Sort by Freq