+++ R/write_object.R
@@ -8,6 +8,9 @@
 #' @param dir a character string specifying the directory (folder) in which save
 #'   the object, defaults to current working directory.
 #' @param date logical, if \code{TRUE} appends the object name with todays date
+#' @param overwrite logical, if \code{TRUE} overwrites existing file of same name
+#' @param row.names logical, if \code{TRUE} adds rownames to output file
+#' @param col.names logical, if \code{TRUE} adds column names to output file
 #' @return Returns nothing. Saves "object_date.tsv" or "object.tsv" to current
 #'   directory.
 #' @importFrom utils write.table
@@ -15,25 +18,32 @@
 #' @examples
 #' write_object(iris2)
 #'
-write_object <- function(object, dir = NULL, date = TRUE) {
+write_object <- function(object, dir = NULL, date = TRUE, overwrite = FALSE, row.names = FALSE, col.names = TRUE) {
   if(date){
     date <- paste0("_", Sys.Date())
   } else {
     date <- ""
   }
-  object <- deparse(substitute(object))
+  name <- deparse(substitute(object))
   if(!is.null(dir)) {
-    file.name <- paste0(dir, "/", object, date, ".tsv")
-    if(file.name %in% list.files(recursive = T)) {
+    file.name <- paste0(dir, "/", name, date, ".tsv")
+    if(file.name %in% list.files(recursive = T) && overwrite == FALSE) {
       stop(paste0("File '", file.name, "' already exists"))
     }
   } else {
-    file.name <- paste0(object, date, ".tsv")
-    if(file.name %in% list.files()) {
+    file.name <- paste0(name, date, ".tsv")
+    if(file.name %in% list.files() && overwrite == FALSE) {
       stop(paste0("File '", file.name, "' already exists"))
     }
   }
+
+  if(col.names){
+    col.names = ifelse(row.names, NA, TRUE)
+  }
+
   write.table(object,
               file = file.name,
-              sep = "\t")
+              sep = "\t",
+              row.names = row.names,
+              col.names = col.names)
 }